requirements.txt
beautifulsoup4==4.12.0
google-generativeai==0.8.4
python-dotenv==1.0.1
requests==2.31.0
python-docx
json